pythonpostgres/requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.1
tenacity>=8.2
openai>=1.37
python-dotenv>=1.0
pytest>=8.0
httpx~=0.28.1